Main parts of interface has been reworked to work with cost as a long type instead of double. New “No category” category implementation.

+++ app/src/main/java/dev/nelson/mot/callback/SetDataFromStatisticLoader.java
@@ -3,5 +3,5 @@ package dev.nelson.mot.callback;
 import java.util.LinkedList;
 
 public interface SetDataFromStatisticLoader {
-    public void setDataFromStatisticLoader(LinkedList<String> categoriesNames, LinkedList<Double> sumPerCategory);
+    public void setDataFromStatisticLoader(LinkedList<String> categoriesNames, LinkedList<Long> sumPerCategory);
 }
